rename ssl_iterator.py
