+++ src/baal/utils/ssl_iterator.py
